feat(P_IMPLEMENTATION_SP_C1_GREEN): green(c1): add extra=forbid to all 50 input models + schema_utils + BaseTool.input_schema
