+++ .st3/state.json
@@ -47,5 +47,5 @@
     }
   ],
   "reconstructed": false,
-  "current_sub_phase": null
+  "current_sub_phase": "red"
 }
