+++ mcp_server/tools/scaffold_artifact.py
@@ -16,7 +16,7 @@ Handles all artifact types (code + documents) via ArtifactManager.
 
 from typing import Any
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, ConfigDict, Field
 
 from mcp_server.core.operation_notes import NoteContext
 from mcp_server.managers.artifact_manager import ArtifactManager
@@ -26,6 +26,8 @@ from mcp_server.tools.tool_result import ToolResult
 
 class ScaffoldArtifactInput(BaseModel):
     """Input for scaffold_artifact tool."""
+
+    model_config = ConfigDict(extra="forbid")
 
     artifact_type: str = Field(
         ..., description="Artifact type ID from registry (e.g., 'dto', 'design', 'worker')"
@@ -56,13 +58,6 @@ class ScaffoldArtifactTool(BranchMutatingTool):
         if manager is None:
             raise ValueError("ArtifactManager must be injected for scaffold_artifact")
         self.manager = manager
-
-    @property
-    def input_schema(self) -> dict[str, Any]:
-        """Return JSON schema for input validation."""
-        if self.args_model is None:
-            return {}
-        return self.args_model.model_json_schema()
 
     async def execute(self, params: ScaffoldArtifactInput, context: NoteContext) -> ToolResult:
         """Execute artifact scaffolding.
